requirements.txt
beautifulsoup4==4.11.1
fake_headers==1.0.2
fastapi==0.89.1
Jinja2==3.1.2
opengraphio==1.0.5
praw==7.6.1
pydantic==1.10.4
python-dotenv==0.21.1
requests==2.28.1
SQLAlchemy==1.4.46
